+++ lib/i18n.dart
@@ -31,12 +31,16 @@ part 'src/i18n/currency_list.dart';
 part 'src/i18n/default_currency_data.dart';
 part 'src/i18n/direction.dart';
 part 'src/i18n/direction_estimator.dart';
+part 'src/i18n/word_count_direction_estimator.dart';
+part 'src/i18n/bidi_policy.dart';
 
+part 'src/i18n/has_direction.dart';
 part 'src/i18n/has_text_direction.dart';
 part 'src/i18n/has_alignment.dart';
 part 'src/i18n/has_horizontal_alignment.dart';
 part 'src/i18n/has_vertical_alignment.dart';
 part 'src/i18n/has_auto_horizontal_alignment.dart';
+part 'src/i18n/auto_direction_handler.dart';
 
 part 'src/i18n/constants/number_constants.dart';