+++ lib/src/ui/integer_box.dart
@@ -6,7 +6,7 @@ part of dart_web_toolkit_ui;
 /**
  * A ValueBox that uses {@link IntegerParser} and {@link IntegerRenderer}.
  */
-class IntBox extends ValueBox<int> {
+class IntegerBox extends ValueBox<int> {
 
-  IntBox() : super.fromElement(new dart_html.TextInputElement(), new IntRenderer.instance(), new IntParser.instance());
+  IntegerBox() : super.fromElement(new dart_html.TextInputElement(), new IntegerRenderer.instance(), new IntegerParser.instance());
 }