+++ lib/text.dart
@@ -20,6 +20,6 @@ part 'src/text/date_time_format_renderer.dart';
 part 'src/text/appendable.dart';
 part 'src/text/double_parser.dart';
 part 'src/text/double_renderer.dart';
-part 'src/text/int_parser.dart';
-part 'src/text/int_renderer.dart';
+part 'src/text/integer_parser.dart';
+part 'src/text/integer_renderer.dart';
 part 'src/text/number_format_renderer.dart';